+++ src/Orleans.sln
@@ -78,7 +78,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Nuget", "Nuget", "{014CD19F
 		NuGet\Microsoft.Orleans.Server.nuspec = NuGet\Microsoft.Orleans.Server.nuspec
 		NuGet\Microsoft.Orleans.Templates.Grains.nuspec = NuGet\Microsoft.Orleans.Templates.Grains.nuspec
 		NuGet\Microsoft.Orleans.Templates.Interfaces.nuspec = NuGet\Microsoft.Orleans.Templates.Interfaces.nuspec
-		NuGet\Microsoft.Orleans.UnitTest.nuspec = NuGet\Microsoft.Orleans.UnitTest.nuspec
+		NuGet\Microsoft.Orleans.TestingHost.nuspec = NuGet\Microsoft.Orleans.TestingHost.nuspec
 		NuGet\Orleans.Grains.props = NuGet\Orleans.Grains.props
 		NuGet\Orleans.Interfaces.props = NuGet\Orleans.Interfaces.props
 		NuGet\ServerInstall.ps1 = NuGet\ServerInstall.ps1
@@ -133,7 +133,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Chocolatey", "Chocolatey", 
 		Chocolatey\orleans.nuspec = Chocolatey\orleans.nuspec
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OrleansUnitTestHost", "OrleansUnitTestHost\OrleansUnitTestHost.csproj", "{40EE3B00-D381-485F-9C69-FF706837DEED}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OrleansTestingHost", "OrleansTestingHost\OrleansTestingHost.csproj", "{40EE3B00-D381-485F-9C69-FF706837DEED}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
