+++ src/OrleansTestingHost/UnitTestUtils.cs
@@ -27,7 +27,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Orleans;
 
-namespace OrleansUnitTestHost
+namespace OrleansTestingHost
 {
     public static class UnitTestUtils
     {
